Add Week 2 Day 3 task for Ayooluwa Oyewo-Use Modules
