+++ submissions/week-2/day-3/Ayooluwa-Oyewo/digital-keycard/src/error.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, PartialEq)]
+pub enum EmployeeError {
+    EmployeeNotFound,
+    EmployeeTerminated,
+    EmployeeNotAllowed,
+}
