+++ submissions/week-2/day-3/Ayooluwa-Oyewo/digital-keycard/src/lib.rs
@@ -1,0 +1,7 @@
+// Core modules
+pub mod employee;
+pub mod error;
+
+// Test module (only available in test configuration)
+#[cfg(test)]
+pub mod tests;
